Added experimental Euler work in HTML.
This is the Continued Fractions chapter of the Introductio.  So far
it seems that MathJax and HTML is a very good, and even convenient,
way to present Euler works.  I also have the start of an English translation
which I am not yet committing.

The resulting text looks very much like the Opera Omnia, which is to
say very much like the original except better in a lot of small ways.