Added temporary figures for long division.
They are simple preformatted blocks.  Mathjax seems to
get in the way of any TeX solution, so a higher quality
figure could be done by generating an svg figure and including
it as an image.

This completes the first pass of the Latin text for Euler's
continued fractions chapter, rendered in html.